+++ Server/Models/Monsters/DoomClaw.cs
@@ -27,7 +27,7 @@ namespace Server.Models.Monsters
             return 0;
         }
 
-        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false,
+        public override long Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false,
             bool canCrit = true, bool canStruck = true)
         {
